Development Build
NOTE: This build is now using the latest ViroReact v2.14.0. This version supports RN 0.59.3. We also upgraded to ARCore 1.7 and was released on April 11th so it's very new. This update was needed to stop some crashes from happening due to the launcher app on Android. It also added support for Animated GIF's, which might come in handy.